cs50p/project/requirements.txt
rich
newsapi-python
tinydb
python-cfonts
passlib
pwinput
python-dotenv